+++ src/components/organisms/AboutModal.tsx
@@ -6,10 +6,10 @@ import DialogTitle from "@mui/material/DialogTitle";
 import DialogContent from "@mui/material/DialogContent";
 import DialogActions from "@mui/material/DialogActions";
 import Divider from "@mui/material/Divider";
-import { useResetAllData } from "@/hooks/useResetAllData";
+import { useResetAllData } from "@/hooks/v1/useResetAllData";
 import { THIS_GITHUB_URL } from "@/constants/meta";
 import { SNS_LIST, TWITTER_URL } from "@/constants/sns";
-import { usePreviewCommand } from "@/hooks/usePreview";
+import { usePreviewCommand } from "@/hooks/v1/usePreview";
 import { PreviewModal } from "@/components/molecules/PreviewModal";
 import { CollaborationsSection as NaiveCollaborationsSection } from "@/components/molecules/CollaborationsSection";
