+++ src/components/organisms/TutorialModal.tsx
@@ -8,7 +8,7 @@ import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
 import DialogTitle from "@mui/material/DialogTitle";
 import DialogContent from "@mui/material/DialogContent";
-import { useStepper } from "@/hooks/useStepper";
+import { useStepper } from "@/hooks/v1/useStepper";
 
 const steps = [
   {
