+++ src/components/pages/ErrorPage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import Button from "@mui/material/Button";
 import { LogoImage } from "@/components/atoms/LogoImage";
-import { useResetAllData } from "@/hooks/useResetAllData";
+import { useResetAllData } from "@/hooks/v1/useResetAllData";
 
 export function ErrorPage() {
   const { reset } = useResetAllData();
