+++ src/components/templates/UserCalendarSection.tsx
@@ -2,8 +2,8 @@ import styled from "styled-components";
 import Paper from "@mui/material/Paper";
 import Box from "@mui/material/Box";
 import Divider from "@mui/material/Divider";
-import { useStoryModal } from "@/hooks/useStoryModal";
-import { useEventModal } from "@/hooks/useEventModal";
+import { useStoryModal } from "@/hooks/v1/useStoryModal";
+import { useEventModal } from "@/hooks/v1/useEventModal";
 import { StoryEditModal } from "@/components/molecules/StoryEditModal";
 import { EventEditModal } from "@/components/molecules/EventEditModal";
 import { UserCalendar } from "@/components/organisms/UserCalendar";
