+++ src/hooks/v1/useResourceGroupLabelContentInTemplateCalendar.tsx
