+++ src/hooks/v1/useResourceGroupLabelContentInUserCalendar.tsx
