+++ src/hooks/v1/useUserCalendarCustomButtons.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import { useUserCalendar } from "@/hooks/useUserCalendar";
-import { useStory } from "@/hooks/useStory";
+import { useUserCalendar } from "@/hooks/v1/useUserCalendar";
+import { useStory } from "@/hooks/v1/useStory";
 
 export const ADD_STORY_BUTTON = "ADD_STORY_BUTTON";
 export const REMOVE_CALENDAR_BUTTON = "REMOVE_CALENDAR_BUTTON";
